Rename cassandra sub-package
